new update

+++ store/models.py
@@ -13,7 +13,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=0)
     description = models.TextField(blank=True, null=True)
     image = models.ImageField(blank=True, null=True)
